Create an event.

This function creates an event script from the provided event template inside a module along with a testtthat test script.

use_event(name, module, with_comments)

Arguments

name
(character(1))
Name of an event to create. This is also be used as its script name.

module
(character(1))
Name of the module folder to add a event file to. The function looks for a folder inside the modules folder at the root folder of the active R project. If the module folder is not found or has not been created this will return an error.

with_comments
(logical(1))
If TRUE the generated event script will contain comments about what each component inside the script does and some recommendations for the user to follow when authoring an event. For advance users, you may not need this hence you may specify FALSE. If missing, it will be prompted in the console for you to decide.

Examples

if (FALSE) {
  # Note: running this will create a folder called "modules" and a sub-folder
  #       to your working directory within the folder called "demography"
  use_module(name = "demography")

  # create an event called 'birth' inside the 'demography' module.
  use_event(name = "birth", module = 'demography')
}